pyproject.toml
[project]
name = "mlops-with-databricks"
version = "0.0.1"
description = "MLOps with Databricks"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "scikit-learn==1.5.2",
    "cloudpickle==3.1.0",
    "mlflow==2.17.0",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "pyarrow==14.0.1",
    "cffi==1.17.1",
    "scipy==1.14.1",
    "matplotlib==3.9.2",
    "databricks-feature-engineering==0.6",
    "pydantic==2.9.2",
    "seaborn>=0.13.2",
    "databricks-feature-lookup==1.2.0",
    "databricks-sdk==0.32.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.3",
       "pre-commit>=4.0.1",
       "databricks-connect>=15.4.1, <16",
       "ipykernel>=6.29.5, <7",
       "ruff>=0.7.0",
       "pip>=24.2"]

[build-system]
requires = [
  "setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 79
indent-width = 4
src = ["src", "test"]
exclude = ["notebooks"]

show-fixes = true

[tool.ruff.format]
indent-style = "space"
exclude = [
  ".git",
  ".github",
  ".idea",
  ".pytest_cache",
  ".ruff_cache",
  ".venv",
  ".vscode",
  "typings/__builtins__.pyi",
  "notebooks",
]

[tool.ruff.lint]
select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
